server/requirements.txt
audioread==3.0.1
blinker==1.8.1
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
coverage==7.5.0
cycler==0.12.1
decorator==5.1.1
exceptiongroup==1.2.1
Flask==3.0.3
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
fonttools==4.51.0
idna==3.7
importlib_metadata==7.1.0
importlib_resources==6.4.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.3
joblib==1.4.0
jsonpickle==3.0.4
kiwisolver==1.4.5
lazy_loader==0.4
librosa==0.10.1
llvmlite==0.42.0
MarkupSafe==2.1.5
matplotlib==3.8.4
mido==1.3.2
more-itertools==10.2.0
msgpack==1.0.8
music21==8.3.0
mysql-connector-python==8.3.0
numba==0.59.1
numpy==1.22.4
packaging==23.2
pillow==10.3.0
platformdirs==4.2.1
pluggy==1.5.0
pooch==1.8.1
pycparser==2.22
pydub==0.25.1
pyparsing==3.1.2
pytest==8.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.31.0
scikit-learn==1.4.2
scipy==1.7.3
six==1.16.0
soundfile==0.12.1
soxr==0.3.7
SQLAlchemy==2.0.29
threadpoolctl==3.5.0
tomli==2.0.1
typing_extensions==4.11.0
urllib3==2.2.1
webcolors==1.13
Werkzeug==3.0.2
zipp==3.18.1
